добавил комменты к методам и классам

+++ src/console/migrations/m190629_112135_user_table_create.php
@@ -17,7 +17,7 @@ class m190629_112135_user_table_create extends Migration
         $this->addTable( [
             'id' => $this->bigPrimaryKey(),
             'username' => $this->string()->notNull()->unique(),
-            'locale' => $this->integer()->notNull(),
+            'locale' => $this->bigInteger()->notNull(),
             'domain_id' => $this->bigInteger(),
             'created_at' => $this->dateTime()->defaultValue(new \yii\db\Expression("NOW()") ),
             'updated_at' => $this->dateTime()->append('ON UPDATE NOW()'),
